+++ changelog.md
@@ -15,6 +15,7 @@
 - Update `vmq_diversity` to newest Luerl version
 - Bumps `rebar3` executable
 - Upgrade Cowboy dependency to 2.8.0
+- Adds support for `auth_source` in MongoDB connections in `vmq_diversity`
 
 ## VerneMQ 1.11.0